#15: Move tests around
